# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Fake flix API",
        "title": "Fake flix API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/content/stream/{id}": {
            "get": {
                "description": "Get stream content",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Video"
                ],
                "summary": "Get stream content",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/video.ResponseSuccess"
                        }
                    }
                }
            }
        },
        "/api/v1/content/upload": {
            "post": {
                "description": "Upload a video file",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Video"
                ],
                "summary": "Upload a video file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Video file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "File name",
                        "name": "filename",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "video.ResponseSuccess": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}